golden_response/requirements.txt
openai>=1.0.0
streamlit>=1.35.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
colorama>=0.4.6
PyPDF2>=3.0.0
python-docx>=1.0.0
pandas>=2.0.0
